chatbot-backend/requirements.txt
fastapi
uvicorn
python-multipart
python-jose[cryptography]
pydantic
llama-cpp-python
requests